Cleaned up loadParty(). Added Group class.
I cleaned up the loadParty() method and modified it to utilize the new
Group class.  The Group class stores Person objects, has a reference to
the Table it is currently at, and knows its size. I also coded the
swapGroups() method. It's still a little basic (there's no way to select
multiple groups to swap in the destination), but it works for now.